+++ tsrc/cli/apply_manifest.py
@@ -1,18 +1,26 @@
-""" Entry point for `tsrc apply-manifest` """
+""" Entry point for `tsrc apply-manifest`. """
 
-import argparse
+from pathlib import Path
+from typing import Any
 
 import cli_ui as ui
-import tsrc.cli
+from argh import arg
+
+import tsrc.manifest
+from tsrc.cli import repos_from_config, workspace_action, workspace_arg
 
 
-def main(args: argparse.Namespace) -> None:
-    workspace = tsrc.cli.get_workspace(args)
-    manifest_path = args.manifest_path
+@workspace_arg  # type: ignore
+@workspace_action
+@arg("manifest_path", help="path to the local manifest", type=Path)  # type: ignore
+def apply_manifest(
+    workspace: tsrc.Workspace, manifest_path: Path, **kwargs: Any
+) -> None:
+    """ apply a local manifest file """
+    ui.info_1("Applying manifest from", manifest_path)
 
-    ui.info_1("Applying manifest from", args.manifest_path)
     manifest = tsrc.manifest.load(manifest_path)
-    workspace.repos = tsrc.cli.repos_from_config(manifest, workspace.config)
+    workspace.repos = repos_from_config(manifest, workspace.config)
     workspace.clone_missing()
     workspace.set_remotes()
     workspace.perform_filesystem_operations()