+++ tsrc/cli/log.py
@@ -1,15 +1,23 @@
-""" Entry point for tsrc log """
+""" Entry point for `tsrc log`. """
 
-import argparse
+from typing import Any
 
 import cli_ui as ui
+from argh import arg
 
 import tsrc
-import tsrc.cli
+from tsrc.cli import repos_action, repos_arg
 
 
-def main(args: argparse.Namespace) -> None:
-    workspace = tsrc.cli.get_workspace_with_repos(args)
+@repos_arg
+@repos_action
+@arg("--from", dest="from", metavar="FROM", help="from ref")  # type: ignore
+@arg("--to", help="to ref")  # type: ignore
+def log(workspace: tsrc.Workspace, **kwargs: Any) -> None:
+    """ show a combined git log for several repositories """
+    from_: str = kwargs["from"]
+    to: str = kwargs["to"] or "HEAD"
+
     all_ok = True
     for repo in workspace.repos:
         full_path = workspace.root_path / repo.dest
@@ -25,7 +33,7 @@ def main(args: argparse.Namespace) -> None:
             "log",
             "--color=always",
             f"--pretty=format:{log_format}",
-            f"{args.from_}...{args.to}",
+            f"{from_}...{to}",
         ]
         rc, out = tsrc.git.run_captured(full_path, *cmd, check=False)
         if rc != 0: