+++ tsrc/cli/status.py
@@ -1,33 +1,47 @@
 """ Entry point for tsrc status """
 
-from typing import Dict, List, Union, Optional, Tuple
-
-import argparse
 import collections
 import shutil
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 import cli_ui as ui
 
 import tsrc
 import tsrc.errors
-import tsrc.cli
 import tsrc.git
+from tsrc.cli import repos_action, repos_arg
 
 
 class ManifestStatus:
+    """ Represent the status of a repo w.r.t the manifest. """
+
     def __init__(self, repo: tsrc.Repo, *, manifest: tsrc.Manifest):
         self.repo = repo
         self.manifest = manifest
         self.incorrect_branch = None  # type: Optional[Tuple[str,str]]
 
     def update(self, git_status: tsrc.git.Status) -> None:
+        """Set self.incorrect_branch if the local git status
+        does not match the branch set in the manifest.
+        """
         expected_branch = self.repo.branch
         actual_branch = git_status.branch
         if actual_branch and actual_branch != expected_branch:
             self.incorrect_branch = (actual_branch, expected_branch)
 
+    def describe(self) -> List[ui.Token]:
+        """ Return a list of tokens suitable for ui.info()`. """
+        res = []  # type: List[ui.Token]
+        incorrect_branch = self.incorrect_branch
+        if incorrect_branch:
+            actual, expected = incorrect_branch
+            res += [ui.red, "(expected: " + expected + ")"]
+        return res
+
 
 class Status:
+    """ Wrapper class for both ManifestStatus and GitStatus"""
+
     def __init__(self, *, git: tsrc.git.Status, manifest: ManifestStatus):
         self.git = git
         self.manifest = manifest
@@ -38,70 +52,14 @@ CollectedStatuses = Dict[str, StatusOrError]
 
 
 def describe_status(status: StatusOrError) -> List[ui.Token]:
-    """ Returns a list of tokens suitable for ui.info() """
+    """ Return a list of tokens suitable for ui.info(). """
     if isinstance(status, tsrc.errors.MissingRepo):
         return [ui.red, "error: missing repo"]
     if isinstance(status, Exception):
         return [ui.red, "error: ", status]
-    return describe_git_status(status.git) + describe_manifest_status(status.manifest)
-
-
-def describe_git_status(git_status: tsrc.git.Status) -> List[ui.Token]:
-    res = []  # type: List[ui.Token]
-    res += describe_branch(git_status)
-    res += describe_position(git_status)
-    res += describe_dirty(git_status)
-    return res
-
-
-def describe_manifest_status(manifest_status: ManifestStatus) -> List[ui.Token]:
-    res = []  # type: List[ui.Token]
-    incorrect_branch = manifest_status.incorrect_branch
-    if incorrect_branch:
-        actual, expected = incorrect_branch
-        res += [ui.red, "(expected: " + expected + ")"]
-
-    return res
-
-
-def describe_branch(git_status: tsrc.git.Status) -> List[ui.Token]:
-    res = []  # type: List[ui.Token]
-    if git_status.branch:
-        res += [ui.green, git_status.branch]
-    elif git_status.sha1:
-        res += [ui.red, git_status.sha1]
-    if git_status.tag:
-        res += [ui.reset, ui.brown, "on", git_status.tag]
-    return res
-
-
-def commit_string(number: int) -> str:
-    if number == 1:
-        return "commit"
-    else:
-        return "commits"
-
-
-def describe_position(git_status: tsrc.git.Status) -> List[ui.Token]:
-    res = []  # type: List[ui.Token]
-    if git_status.ahead != 0:
-        up = ui.Symbol("↑", "+")
-        n_commits = commit_string(git_status.ahead)
-        ahead_desc = f"{up.as_string}{git_status.ahead} {n_commits}"
-        res += [ui.blue, ahead_desc, ui.reset]
-    if git_status.behind != 0:
-        down = ui.Symbol("↓", "-")
-        n_commits = commit_string(git_status.behind)
-        behind_desc = f"{down.as_string}{git_status.behind} {n_commits}"
-        res += [ui.blue, behind_desc, ui.reset]
-    return res
-
-
-def describe_dirty(git_status: tsrc.git.Status) -> List[ui.Token]:
-    res = []  # type: List[ui.Token]
-    if git_status.dirty:
-        res += [ui.red, "(dirty)"]
-    return res
+    git_status = status.git.describe()
+    manifest_status = status.manifest.describe()
+    return git_status + manifest_status
 
 
 def erase_last_line() -> None:
@@ -110,6 +68,10 @@ def erase_last_line() -> None:
 
 
 class StatusCollector(tsrc.Task[tsrc.Repo]):
+    """Implement a Task to collect local git status and
+    stats w.r.t the manifest for each repo.
+    """
+
     def __init__(self, workspace: tsrc.Workspace) -> None:
         self.workspace = workspace
         self.manifest = workspace.get_manifest()
@@ -154,7 +116,9 @@ class StatusCollector(tsrc.Task[tsrc.Repo]):
             ui.info(*message)
 
 
-def main(args: argparse.Namespace) -> None:
-    workspace = tsrc.cli.get_workspace_with_repos(args)
+@repos_arg
+@repos_action
+def status(workspace: tsrc.Workspace, **kwargs: Any) -> None:
+    """ display workspace status summary """
     status_collector = StatusCollector(workspace)
     tsrc.run_sequence(workspace.repos, status_collector)