+++ tsrc/git.py
@@ -3,12 +3,15 @@
 
 import os
 import subprocess
-from typing import Any, Dict, Iterable, Tuple, Optional  # noqa
+from pathlib import Path
+from typing import Any, Dict, Iterable, List, Optional, Tuple  # noqa
 
-from path import Path
 import cli_ui as ui
 
 import tsrc
+
+UP = ui.Symbol("↑", "+").as_string
+DOWN = ui.Symbol("↓", "-").as_string
 
 
 class Error(tsrc.Error):
@@ -45,7 +48,17 @@ def assert_working_path(path: Path) -> None:
 
 
 class Status:
+    """Represent a status of a git repo.
+
+    Usage:
+    >>> status = Status(repo_path)
+    >>> status.update()
+    """
+
     def __init__(self, working_path: Path) -> None:
+        # Note: at this point no information is known, and all
+        # attributes have their default value.
+        self.empty = False
         self.working_path = working_path
         self.untracked = 0
         self.staged = 0
@@ -59,7 +72,13 @@ class Status:
         self.sha1 = None  # type: Optional[str]
 
     def update(self) -> None:
-        self.update_sha1()
+        # Try and gather as many information about the git repository as
+        # possible.
+        try:
+            self.update_sha1()
+        except CommandError:
+            self.empty = True
+            return
         self.update_branch()
         self.update_tag()
         self.update_remote_status()
@@ -110,9 +129,62 @@ class Status:
                 self.added += 1
                 self.dirty = True
 
+    def describe(self) -> List[ui.Token]:
+        """ Return a list of tokens suitable for ui.info. """
+        res = []  # type: List[ui.Token]
+        if self.empty:
+            return [ui.red, "empty"]
+        res += self.describe_branch()
+        res += self.describe_position()
+        res += self.describe_dirty()
+        return res
+
+    def describe_branch(self) -> List[ui.Token]:
+        res = []  # type: List[ui.Token]
+        if self.branch:
+            res += [ui.green, self.branch, ui.reset]
+        elif self.sha1:
+            res += [ui.red, self.sha1, ui.reset]
+        if self.tag:
+            res += [ui.brown, "on", self.tag, ui.reset]
+        return res
+
+    @staticmethod
+    def commit_string(number: int) -> str:
+        """ Describe the number of commit with correct pluralization. """
+
+        if number == 1:
+            return "commit"
+        else:
+            return "commits"
+
+    def describe_position(self) -> List[ui.Token]:
+        """Return a status looking like `↑2↓1` if the branch
+        is 2 commits ahead and one commit behind its upstream,
+        as a list of tokens suitable for `ui.info()`.
+
+        """
+        res = []  # type: List[ui.Token]
+        if self.ahead != 0:
+            n_commits = Status.commit_string(self.ahead)
+            ahead_desc = f"{UP}{self.ahead} {n_commits}"
+            res += [ui.blue, ahead_desc, ui.reset]
+        if self.behind != 0:
+            n_commits = Status.commit_string(self.behind)
+            behind_desc = f"{DOWN}{self.behind} {n_commits}"
+            res += [ui.blue, behind_desc, ui.reset]
+        return res
+
+    def describe_dirty(self) -> List[ui.Token]:
+        """ Add the `(dirty)` colored string if the repo is dirty. """
+        res = []  # type: List[ui.Token]
+        if self.dirty:
+            res += [ui.red, "(dirty)", ui.reset]
+        return res
+
 
 def run(working_path: Path, *cmd: str, check: bool = True) -> None:
-    """ Run git `cmd` in given `working_path`
+    """Run git `cmd` in given `working_path`.
 
     Raise GitCommandError if return code is non-zero and `check` is True.
     """
@@ -127,11 +199,11 @@ def run(working_path: Path, *cmd: str, check: bool = True) -> None:
 
 
 def run_captured(working_path: Path, *cmd: str, check: bool = True) -> Tuple[int, str]:
-    """ Run git `cmd` in given `working_path`, capturing the output
+    """Run git `cmd` in given `working_path`, capturing the output.
 
     Return a tuple (returncode, output).
 
-    Raise GitCommandError if return code is non-zero and check is True
+    Raise GitCommandError if return code is non-zero and check is True.
     """
     assert_working_path(working_path)
     git_cmd = list(cmd)