+++ tsrc/workspace/file_system_operator.py
@@ -1,13 +1,18 @@
+from pathlib import Path
 from typing import List
 
 import cli_ui as ui
-from path import Path
 
 import tsrc
 import tsrc.executor
 
 
 class FileSystemOperator(tsrc.executor.Task[tsrc.FileSystemOperation]):
+    """Implement file system operations to be run once every missing
+    repo has been cloned, like copying files or creating symlinks.
+
+    """
+
     def __init__(self, workspace_path: Path, repos: List[tsrc.Repo]) -> None:
         self.workspace_path = workspace_path
         self.repos = repos